+++ android/src/main/java/com/bitmovin/player/reactnative/RNPlayerViewPackage.kt
@@ -24,10 +24,11 @@ class RNPlayerViewPackage : ReactPackage {
             PlayerModule(reactContext),
             SourceModule(reactContext),
             DrmModule(reactContext),
-            AnalyticsModule(reactContext),
+            PlayerAnalyticsModule(reactContext),
             RNPlayerViewManager(reactContext),
             FullscreenHandlerModule(reactContext),
-            CustomMessageHandlerModule(reactContext)
+            CustomMessageHandlerModule(reactContext),
+            BitmovinCastManagerModule(reactContext),
         )
     }