+++ src/drm/index.ts
@@ -14,11 +14,15 @@ const DrmModule = NativeModules.DrmModule;
  */
 export interface DrmConfig extends NativeInstanceConfig {
   /**
-   * FairPlay specific configuration. Only applicable for iOS.
+   * FairPlay specific configuration.
+   *
+   * @platform iOS
    */
   fairplay?: FairplayConfig;
   /**
-   * Widevine specific configuration. Only applicable for Android.
+   * Widevine specific configuration.
+   *
+   * @platform Android, iOS (only for casting).
    */
   widevine?: WidevineConfig;
 }